requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==4.2.11
django-crispy-forms==2.1
django-environ==0.11.2
django-widget-tweaks==1.5.0
idna==3.7
install==1.3.5
pillow==10.3.0
psycopg2==2.9.9
requests==2.32.2
SpeechRecognition==3.10.4
sqlparse==0.5.0
typing-extensions==4.11.0
urllib3==2.2.1
